Update README.md
# Python-R-Multi-Omics-packages-tutorial
This repository provides a comprehensive list of packages and online tutorials for Multi-Omics and NGS data analysis in both R and Python programming languages.

This list covers a comprehensive set of tutorial, Jupyter, Github online textbooks and repositories that covers the following types of Omics data:


1-Pathway Enrichment Analysis.
2-Gene Set Enrichment Analysis.
3-Metabolomics using R.
4-SNP, GWAS, and Population Genetics.
5-MicroArray using R/Python.
6-MiRNA.
7-Chip-seq R/Python.
8-Methylation R/Python.
9-ATAC R/Python.
10-Differential Analysis R/Python.
11-RNA-Seq R/Python.
12-Multi-omics R/Python.
13-Single-Cell R/Python.
14-Transcriptomics analysis.
15-Spatial transcriptomics .
16-NGS.
17-Proteomics.
18-MetaGenomics.
19-computational genomics .
20-Python Multi-Omics Packages Articles.
21-Bioinformatics Courses.
22-Galaxy.
23-Comprehensive guides.
24-Activity based protein differential expression analysis using R.
25-Genomics data science john hopkins.
26-intro to r system biology.
27-Exploratory analysis of transcriptomics data in R.
28-Systems Biology and Biotechnology Specialization.

To access the links go to the file named "Links for online multi-omics textbooks tutorials and packages" in the list of files above.

Soon other types of resoruces like Video tutorials, textbooks, and code tutorials will be added in a separate files.